Change %d to %ld and cast to long, which is safer than casting 64-bit values to 32-bit int. Fix some int/magma_int_t compatability. (in src)
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@9363 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ src/dgeev_m.cpp
@@ -507,10 +507,10 @@ CLEANUP:
     
     timer_stop( time_total );
     flops_stop( flop_total );
-    timer_printf( "dgeev times n %5d, gehrd %7.3f, unghr %7.3f, hseqr %7.3f, trevc %7.3f, total %7.3f, sum %7.3f\n",
-                  (int) n, time_gehrd, time_unghr, time_hseqr, time_trevc, time_total, time_sum );
-    timer_printf( "dgeev flops n %5d, gehrd %7lld, unghr %7lld, hseqr %7lld, trevc %7lld, total %7lld, sum %7lld\n",
-                  (int) n, flop_gehrd, flop_unghr, flop_hseqr, flop_trevc, flop_total, flop_sum );
+    timer_printf( "dgeev times n %5ld, gehrd %7.3f, unghr %7.3f, hseqr %7.3f, trevc %7.3f, total %7.3f, sum %7.3f\n",
+                  long(n), time_gehrd, time_unghr, time_hseqr, time_trevc, time_total, time_sum );
+    timer_printf( "dgeev flops n %5ld, gehrd %7lld, unghr %7lld, hseqr %7lld, trevc %7lld, total %7lld, sum %7lld\n",
+                  long(n), flop_gehrd, flop_unghr, flop_hseqr, flop_trevc, flop_total, flop_sum );
     
     work[0] = magma_dmake_lwork( optwrk );
